requirements.txt
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-resized==0.3.11
django-rest-knox==4.1.0
django-summernote==0.8.11.6
django-taggit==1.3.0 
djangorestframework==3.12.2
drf-yasg==1.20.0
Pillow==8.0.1
six==1.15.0

gunicorn==20.1.0
psycopg2-binary==2.9.1
django-storages==1.11.1
boto3==1.17.96